requirements/base.txt
# SHA1:4857d451650525117a19fae346a502c6db411ced
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
attrs==19.3.0
certifi==2019.11.28       # via python-telegram-bot, requests
cffi==1.13.2              # via cryptography
chardet==3.0.4            # via requests
cryptography==2.8         # via python-telegram-bot
future==0.18.2            # via python-telegram-bot
gitdb2==2.0.6             # via gitpython
gitpython==3.0.5
idna==2.8                 # via requests
pycparser==2.19           # via cffi
python-telegram-bot==12.3.0
pytz==2019.3              # via tzlocal
requests==2.22.0
six==1.14.0               # via cryptography
smmap2==2.0.5             # via gitdb2
tornado==6.0.3            # via python-telegram-bot
tzlocal==2.0.0
urllib3==1.25.7           # via requests
